+++ doc/source/changelog.rst
@@ -10,6 +10,8 @@ Version 1.8.0
 New:
 
   * A method to retrieve zone data has been added.
+  * A method to retrieve Z, A, and state label from a nuclide name
+    has been added.
 
 Version 1.7.1
 -------------